Interation02 thai

See merge request summer20241/summer251/se1941-jv/g1/personalized-learning-system!31

+++ src/main/java/vn/fu_ohayo/service/JwtService.java
@@ -7,7 +7,7 @@ import vn.fu_ohayo.enums.TokenType;
 import java.util.Collection;
 
 public interface JwtService {
-    String generateAccessToken(int userId, String email, Collection<? extends GrantedAuthority> authorities);
-    String generateRefreshToken(int userId, String email, Collection<? extends GrantedAuthority> authorities);
+    String generateAccessToken(Long userId, String email, Collection<? extends GrantedAuthority> authorities);
+    String generateRefreshToken(Long userId, String email, Collection<? extends GrantedAuthority> authorities);
     ExtractTokenResponse extractUserInformation(String token, TokenType type);
 }
